implemented add products function.

+++ src/models/CPU.java
@@ -2,7 +2,7 @@ package models;
 
 public class CPU extends Product {
 	private int CoresNb;
-	private int Freq;
+	private Float Freq;
 	private int Speed;
 	public int getCoresNb() {
 		return CoresNb;
@@ -10,10 +10,10 @@ public class CPU extends Product {
 	public void setCoresNb(int coresNb) {
 		CoresNb = coresNb;
 	}
-	public int getFreq() {
+	public Float getFreq() {
 		return Freq;
 	}
-	public void setFreq(int freq) {
+	public void setFreq(Float freq) {
 		Freq = freq;
 	}
 	public int getSpeed() {
